requirements.txt
pandas
numpy
streamlit
fastapi
uvicorn
langchain>=0.3.0
langchain-community
langchain-chroma
langchain-huggingface
pypdf
python-docx
python-pptx
pandas
markdown
unstructured
setuptools

sentence-transformers
python-dotenv
chromadb

python-multipart
requests
-e .